# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trading Signal Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: signals
    description: Trading signal operations
  - name: bot
    description: Trading bot operations
  - name: market
    description: Market status and spread analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /signals/webhook:
    post:
      operationId: receiveWebhook
      tags: [signals]
      summary: Receive TradingView alert webhook
      description: Receives a JSON alert payload from TradingView and stores it as a signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookPayload"
      responses:
        "201":
          description: Signal stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"

  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List all signals
      description: Returns paginated list of signals, newest first
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: symbol
          in: query
          schema:
            type: string
        - name: dir
          in: query
          schema:
            type: string
            enum: [LONG, SHORT]
        - name: triggered
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalList"

  /signals/latest:
    get:
      operationId: getLatestSignal
      tags: [signals]
      summary: Get most recent signal
      responses:
        "200":
          description: Latest signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "404":
          description: No signals yet

  /signals/stats:
    get:
      operationId: getSignalStats
      tags: [signals]
      summary: Aggregate stats across all signals
      responses:
        "200":
          description: Signal statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalStats"

  /bot/config:
    get:
      operationId: getBotConfig
      tags: [bot]
      summary: Get bot configuration
      responses:
        "200":
          description: Bot configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
    put:
      operationId: updateBotConfig
      tags: [bot]
      summary: Update bot configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfigPatch"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"

  /bot/status:
    get:
      operationId: getBotStatus
      tags: [bot]
      summary: Get bot runtime status including balance
      responses:
        "200":
          description: Bot status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"

  /bot/performance:
    get:
      operationId: getBotPerformance
      tags: [bot]
      summary: Get bot performance statistics
      responses:
        "200":
          description: Performance stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotPerformance"

  /bot/test-signal:
    post:
      operationId: sendTestSignal
      tags: [bot]
      summary: Fire a simulated test signal using live market prices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestSignalRequest"
      responses:
        "201":
          description: Test signal created and queued for execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"

  /bot/trades:
    get:
      operationId: listTrades
      tags: [bot]
      summary: List all trades
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeList"

  /bot/equity-curve:
    get:
      operationId: getEquityCurve
      tags: [bot]
      summary: Closed trade equity curve
      responses:
        "200":
          description: Equity curve data points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquityCurve"

  /bot/trades/by-signal/{signalId}:
    get:
      operationId: getTradeBySignal
      tags: [bot]
      summary: Get the trade linked to a signal (if any)
      parameters:
        - name: signalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: No trade for this signal

  /bot/trades/{id}:
    get:
      operationId: getTrade
      tags: [bot]
      summary: Get a single trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
    delete:
      operationId: cancelTrade
      tags: [bot]
      summary: Cancel an open trade (cancels SL/TP orders)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"

  /market/status:
    get:
      operationId: getMarketStatus
      tags: [market]
      summary: Get live market regime, dominance spreads, and rotation state
      responses:
        "200":
          description: Market status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketStatus"

  /signals/{id}:
    get:
      operationId: getSignal
      tags: [signals]
      summary: Get a single signal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "404":
          description: Not found
    delete:
      operationId: deleteSignal
      tags: [signals]
      summary: Delete a signal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    WebhookPayload:
      type: object
      description: JSON alert payload from TradingView Pine Script
      properties:
        symbol:
          type: string
        tf:
          type: string
        mode:
          type: string
        macro_tf:
          type: string
        dir:
          type: string
          enum: [LONG, SHORT]
        auth:
          type: string
        pa_mss:
          type: number
        ob_low:
          type: number
        ob_high:
          type: number
        zone_low:
          type: number
        zone_high:
          type: number
        entry_ref:
          type: number
        sl:
          type: number
        rr1:
          type: number
        tp1:
          type: number
        tp2:
          type: number
        tp3:
          type: number
        conf:
          type: integer
        conf_w:
          type: number
        risk_tier:
          type: string
          enum: [HIGH_CONF, MID_CONF, LOW_CONF]
        grade:
          type: string
        rot:
          type: string
        rot_score:
          type: number
        block_reason:
          type: string
        tf2:
          type: string
        tf3:
          type: string
      required:
        - symbol
        - tf
        - dir

    Signal:
      type: object
      properties:
        id:
          type: integer
        receivedAt:
          type: string
          format: date-time
        symbol:
          type: string
        tf:
          type: string
        mode:
          type: string
        macroTf:
          type: string
        dir:
          type: string
          enum: [LONG, SHORT]
        auth:
          type: string
        paMss:
          type: number
          nullable: true
        obLow:
          type: number
          nullable: true
        obHigh:
          type: number
          nullable: true
        zoneLow:
          type: number
          nullable: true
        zoneHigh:
          type: number
          nullable: true
        entryRef:
          type: number
          nullable: true
        sl:
          type: number
          nullable: true
        rr1:
          type: number
          nullable: true
        tp1:
          type: number
          nullable: true
        tp2:
          type: number
          nullable: true
        tp3:
          type: number
          nullable: true
        conf:
          type: integer
          nullable: true
        confW:
          type: number
          nullable: true
        riskTier:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
        triggered:
          type: boolean
        rot:
          type: string
          nullable: true
        rotScore:
          type: number
          nullable: true
        blockReason:
          type: string
          nullable: true
        tf2:
          type: string
          nullable: true
        tf3:
          type: string
          nullable: true
        rawPayload:
          type: object
          nullable: true
      required:
        - id
        - receivedAt
        - symbol
        - tf
        - dir
        - triggered

    SignalList:
      type: object
      properties:
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - signals
        - total
        - limit
        - offset

    SignalStats:
      type: object
      properties:
        total:
          type: integer
        triggered:
          type: integer
        blocked:
          type: integer
        longs:
          type: integer
        shorts:
          type: integer
        avgConfW:
          type: number
          nullable: true
        highConf:
          type: integer
        midConf:
          type: integer
        lowConf:
          type: integer
        symbolBreakdown:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              count:
                type: integer
            required:
              - symbol
              - count
        gradeBreakdown:
          type: array
          items:
            type: object
            properties:
              grade:
                type: string
              count:
                type: integer
            required:
              - grade
              - count
      required:
        - total
        - triggered
        - blocked
        - longs
        - shorts
        - highConf
        - midConf
        - lowConf
        - symbolBreakdown
        - gradeBreakdown

    EquityCurvePoint:
      type: object
      properties:
        date:
          type: string
          format: date-time
        pnl:
          type: number
        cumulative:
          type: number
        symbol:
          type: string
        closeReason:
          type: string
          nullable: true
        returnPct:
          type: number
          nullable: true
      required:
        - date
        - pnl
        - cumulative
        - symbol

    EquityCurve:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/EquityCurvePoint"
        totalPnl:
          type: number
        totalReturnPct:
          type: number
          nullable: true
      required:
        - points
        - totalPnl

    BotConfig:
      type: object
      properties:
        id:
          type: integer
        enabled:
          type: boolean
        paperMode:
          type: boolean
        longOnly:
          type: boolean
        tradingMode:
          type: string
          description: "all | scalp | intraday | swing | position"
        compoundingEnabled:
          type: boolean
        compoundBalance:
          type: number
          nullable: true
        positionSizeUsdt:
          type: number
        minConfW:
          type: number
        minGrade:
          type: string
        allowedSymbols:
          type: string
        maxOpenTrades:
          type: integer
        cooldownMinutes:
          type: integer
        slEnabled:
          type: boolean
        tp1Enabled:
          type: boolean
        tp2Enabled:
          type: boolean
        tp3Enabled:
          type: boolean
        tp1Pct:
          type: number
        tp2Pct:
          type: number
        tp3Pct:
          type: number
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - enabled
        - paperMode
        - positionSizeUsdt
        - minConfW
        - minGrade
        - allowedSymbols
        - maxOpenTrades
        - cooldownMinutes
        - longOnly
        - tradingMode
        - compoundingEnabled
        - slEnabled
        - tp1Enabled
        - tp2Enabled
        - tp3Enabled
        - tp1Pct
        - tp2Pct
        - tp3Pct
        - updatedAt

    BotConfigPatch:
      type: object
      properties:
        enabled:
          type: boolean
        paperMode:
          type: boolean
        longOnly:
          type: boolean
        tradingMode:
          type: string
        compoundingEnabled:
          type: boolean
        compoundBalance:
          type: number
          nullable: true
        positionSizeUsdt:
          type: number
        minConfW:
          type: number
        minGrade:
          type: string
        allowedSymbols:
          type: string
        maxOpenTrades:
          type: integer
        cooldownMinutes:
          type: integer
        slEnabled:
          type: boolean
        tp1Enabled:
          type: boolean
        tp2Enabled:
          type: boolean
        tp3Enabled:
          type: boolean
        tp1Pct:
          type: number
        tp2Pct:
          type: number
        tp3Pct:
          type: number

    BotStatus:
      type: object
      properties:
        enabled:
          type: boolean
        paperMode:
          type: boolean
        usdtBalance:
          type: number
          nullable: true
        openTrades:
          type: integer
        totalTrades:
          type: integer
        apiConfigured:
          type: boolean
        lastSyncAt:
          type: string
          format: date-time
          nullable: true
      required:
        - enabled
        - paperMode
        - openTrades
        - totalTrades
        - apiConfigured

    BotPerformance:
      type: object
      properties:
        totalTrades:
          type: integer
        closedTrades:
          type: integer
        openTrades:
          type: integer
        paperTrades:
          type: integer
        winRate:
          type: number
          nullable: true
        totalPnl:
          type: number
        avgPnl:
          type: number
          nullable: true
        bestTrade:
          type: number
          nullable: true
        worstTrade:
          type: number
          nullable: true
        longWins:
          type: integer
        longLosses:
          type: integer
        shortWins:
          type: integer
        shortLosses:
          type: integer
      required:
        - totalTrades
        - closedTrades
        - openTrades
        - paperTrades
        - totalPnl
        - longWins
        - longLosses
        - shortWins
        - shortLosses

    TestSignalRequest:
      type: object
      properties:
        symbol:
          type: string
          description: Trading pair e.g. BTCUSDT
        dir:
          type: string
          enum: [LONG, SHORT]
        tf:
          type: string
          description: Timeframe e.g. 4H
        grade:
          type: string
        confW:
          type: number
      required:
        - symbol
        - dir

    Trade:
      type: object
      properties:
        id:
          type: integer
        signalId:
          type: integer
          nullable: true
        symbol:
          type: string
        gateSymbol:
          type: string
        side:
          type: string
        status:
          type: string
        positionSizeUsdt:
          type: number
          nullable: true
        quantity:
          type: number
          nullable: true
        entryPrice:
          type: number
          nullable: true
        livePrice:
          type: number
          nullable: true
        entryOrderId:
          type: string
          nullable: true
        slOrderId:
          type: string
          nullable: true
        tp1OrderId:
          type: string
          nullable: true
        tp2OrderId:
          type: string
          nullable: true
        tp3OrderId:
          type: string
          nullable: true
        slPrice:
          type: number
          nullable: true
        tp1Price:
          type: number
          nullable: true
        tp2Price:
          type: number
          nullable: true
        tp3Price:
          type: number
          nullable: true
        closePrice:
          type: number
          nullable: true
        closeReason:
          type: string
          nullable: true
        pnl:
          type: number
          nullable: true
        errorMessage:
          type: string
          nullable: true
        paperMode:
          type: boolean
        createdAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - symbol
        - gateSymbol
        - side
        - status
        - paperMode
        - createdAt

    TradeList:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
      required:
        - trades
        - total

    MarketSpreads:
      type: object
      properties:
        btcDivStable:
          type: number
          description: "BTC.D / (USDT.D+USDC.D)"
        stableDivBtc:
          type: number
          description: "(USDT.D+USDC.D) / BTC.D"
        btcMinusStable:
          type: number
          description: "BTC.D - Stable.D"
        largecapDivOthers:
          type: number
          description: "(BTC.D+ETH.D) / OTHERS.D"
        totalDivStable:
          type: number
          description: "TOTALES.D / Stable.D"
        usdtDivUsdc:
          type: number
          description: "USDT.D / USDC.D"
      required:
        - btcDivStable
        - stableDivBtc
        - btcMinusStable
        - largecapDivOthers
        - totalDivStable
        - usdtDivUsdc

    MarketStatus:
      type: object
      properties:
        btcDominance:
          type: number
        ethDominance:
          type: number
        stableDominance:
          type: number
        othersDominance:
          type: number
        totalMarketCapUsd:
          type: number
          nullable: true
        marketCapChange24h:
          type: number
          nullable: true
        spreads:
          $ref: "#/components/schemas/MarketSpreads"
        regime:
          type: string
          enum: [RISK_OFF, BTC_DOMINANT, ETH_SEASON, ALT_SEASON, NEUTRAL]
        regimeScore:
          type: number
        latestRotation:
          type: string
          nullable: true
        latestRotScore:
          type: number
          nullable: true
        lastUpdated:
          type: string
          format: date-time
      required:
        - btcDominance
        - ethDominance
        - stableDominance
        - othersDominance
        - spreads
        - regime
        - regimeScore
        - lastUpdated