# lib/api-zod/src/generated/types/botConfigPatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfigPatch {
  enabled?: boolean;
  paperMode?: boolean;
  longOnly?: boolean;
  tradingMode?: string;
  compoundingEnabled?: boolean;
  compoundBalance?: number | null;
  positionSizeUsdt?: number;
  minConfW?: number;
  minGrade?: string;
  allowedSymbols?: string;
  maxOpenTrades?: number;
  cooldownMinutes?: number;
  slEnabled?: boolean;
  tp1Enabled?: boolean;
  tp2Enabled?: boolean;
  tp3Enabled?: boolean;
  tp1Pct?: number;
  tp2Pct?: number;
  tp3Pct?: number;
}