# lib/api-zod/src/generated/types/marketSpreads.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketSpreads {
  /** BTC.D / (USDT.D+USDC.D) */
  btcDivStable: number;
  /** (USDT.D+USDC.D) / BTC.D */
  stableDivBtc: number;
  /** BTC.D - Stable.D */
  btcMinusStable: number;
  /** (BTC.D+ETH.D) / OTHERS.D */
  largecapDivOthers: number;
  /** TOTALES.D / Stable.D */
  totalDivStable: number;
  /** USDT.D / USDC.D */
  usdtDivUsdc: number;
}