# lib/api-zod/src/generated/types/marketStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Signal Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketSpreads } from "./marketSpreads";
import type { MarketStatusRegime } from "./marketStatusRegime";

export interface MarketStatus {
  btcDominance: number;
  ethDominance: number;
  stableDominance: number;
  othersDominance: number;
  totalMarketCapUsd?: number | null;
  marketCapChange24h?: number | null;
  spreads: MarketSpreads;
  regime: MarketStatusRegime;
  regimeScore: number;
  latestRotation?: string | null;
  latestRotScore?: number | null;
  lastUpdated: Date;
}